requirements.txt
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Config & Validation
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database (async MySQL)
sqlalchemy[asyncio]>=2.0.0
asyncmy>=0.2.9

# Auth (bcrypt 4.3.0 dùng trực tiếp, không qua passlib)
python-jose[cryptography]>=3.3.0
bcrypt==4.3.0

# HTTP Client (tests)
httpx>=0.26.0

# Dev & Test
pytest>=7.4.0
pytest-asyncio>=0.23.0
